+++ src/constants/airspaceOLD/lowMoas.js
@@ -1,7 +1,0 @@
-import { arizona } from './lowMoas/arizona'
-import { colorado } from './lowMoas/colorado'
-import { georgia } from './lowMoas/georgia'
-import { germany } from './lowMoas/germany'
-import { southKorea } from './lowMoas/southKorea'
-
-export const lowMoas = arizona.concat(colorado, georgia, germany, southKorea)